requirements.txt
python-dotenv
llama-index
llama-parse
nest-asyncio
duckdb
llama-index-vector-stores-duckdb